+++ react/shared/components/IconSelector/index.tsx
@@ -2,6 +2,7 @@ import type { FC } from 'react'
 import React, { useMemo } from 'react'
 import { Icon } from 'vtex.store-icons'
 import { EXPERIMENTAL_Select as VtexSelect } from 'vtex.styleguide'
+import { FormattedMessage } from 'react-intl'
 
 import useStoreIconPack from './useStoreIconPack'
 
@@ -16,7 +17,6 @@ export interface IconSelectorProps {
 
 export const IconSelector: FC<IconSelectorProps> = ({ onChange }) => {
   const { IconPack, iconPackList } = useStoreIconPack()
-
   const options: Option[] = useMemo(() => {
     return iconPackList.map((icon) => ({
       label: (
@@ -39,6 +39,9 @@ export const IconSelector: FC<IconSelectorProps> = ({ onChange }) => {
         multi={false}
         options={options}
         onChange={onChange}
+        placeholder={
+          <FormattedMessage id="admin/mega-menu.items.input3Form.placeholder" />
+        }
       />
     </div>
   )