+++ react/shared/messages.ts
@@ -140,6 +140,18 @@ export const messagesFirstLevel = defineMessages({
     defaultMessage: '',
     id: 'admin/mega-menu.items.fourColumn',
   },
+  fiveColumn: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.fiveColumn',
+  },
+  emptyTitle: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.emptyTitle',
+  },
+  empty: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.empty',
+  },
   titleModal: {
     defaultMessage: '',
     id: 'admin/mega-menu.items.titleModal',
@@ -194,6 +206,18 @@ export const messageSecondLevel = defineMessages({
   fourColumn: {
     defaultMessage: '',
     id: 'admin/mega-menu.items.fourColumn',
+  },
+  fiveColumn: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.fiveColumn',
+  },
+  emptyTitle: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.emptyTitle',
+  },
+  empty: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.empty',
   },
   titleModal: {
     defaultMessage: '',
@@ -254,6 +278,18 @@ export const messageThirdLevel = defineMessages({
     defaultMessage: '',
     id: 'admin/mega-menu.items.fourColumn',
   },
+  fiveColumn: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.fiveColumn',
+  },
+  emptyTitle: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.emptyTitle',
+  },
+  empty: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.items.empty',
+  },
   titleModal: {
     defaultMessage: '',
     id: 'admin/mega-menu.items.titleModal',
@@ -296,5 +332,32 @@ export const MessagesBackupData = defineMessages({
   duplicateData: {
     defaultMessage: '',
     id: 'admin/mega-menu.backup.duplicateData',
+  },
+})
+
+export const messageSettings = defineMessages({
+  caution: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.settings.caution',
+  },
+  orientationHV: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.settings.orientationHV',
+  },
+  horizontal: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.settings.horizontal',
+  },
+  vertical: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.settings.vertical',
+  },
+  close: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.settings.close',
+  },
+  confirm: {
+    defaultMessage: '',
+    id: 'admin/mega-menu.settings.confirm',
   },
 })